==> notebooks/counterfactuals/01-training.ipynb <==
# Model training

import warnings
warnings.filterwarnings('ignore')

## Data

import pandas as pd

df = pd.read_csv("../../datasets/FICO/heloc_dataset_v1.csv")

TARGET = 'RiskPerformance'

df[TARGET] = df[TARGET].factorize()[0]

df.describe().T

from sklearn.model_selection import train_test_split

train, test = train_test_split(df, test_size=0.25, random_state=42)

train_x = train[df.columns[~train.columns.isin([TARGET])]]
train_y = train[TARGET]

test_x = test[df.columns[~test.columns.isin([TARGET])]]
test_y = test[TARGET]

## XGBoost model

from sklearn.model_selection import GridSearchCV
from xgboost.sklearn import XGBClassifier
from typing import Tuple


def find_best_xgboost_model(
    train_x: pd.DataFrame, train_y: pd.Series
) -> Tuple[dict, float]:
    scale_pos_weight = (len(train_y) - train_y.sum()) / train_y.sum()

    param_test = {
        "max_depth": [2, 4, 6, 8],
        "min_child_weight": [1, 2, 3, 4],
        "learning_rate": [0.1],
        "max_delta_step": [0.4, 0.6, 0.8, 1.0],
        "subsample": [0.9, 0.95, 1.0],
        "colsample_bytree": [0.9, 0.95, 1.0],
        "n_estimators": [10, 100, 200, 300],
    }

    gsearch = GridSearchCV(
        estimator=XGBClassifier(
            use_label_encoder=False,
            verbosity=0,
            silent=True,
            objective="reg:logistic",
            scale_pos_weight=scale_pos_weight,
            seed=27,
        ),
        param_grid=param_test,
        scoring="roc_auc",
        n_jobs=-1,
        cv=8,
    )

    gsearch.fit(train_x, train_y)

    return gsearch.best_params_, gsearch.best_score_

best_params, best_score = find_best_xgboost_model(train_x, train_y)

best_params

best_score

scale_pos_weight = (len(train_y) - train_y.sum()) / train_y.sum()

model = XGBClassifier(objective='reg:logistic',
                              seed=27,
        min_child_weight= best_params['min_child_weight'],
        max_delta_step= best_params['max_delta_step'],
        subsample= best_params['subsample'],
        colsample_bytree= best_params['colsample_bytree'],
                              max_depth=best_params['max_depth'],
                              learning_rate=best_params['learning_rate'],
                              n_estimators=best_params['n_estimators'],
                              scale_pos_weight=scale_pos_weight,
                              )

model.fit(train_x, train_y)

model.save_model("../../models/xgboost.json")

## MLP/Adam model

from sklearn.neural_network import MLPClassifier

from sklearn.model_selection import GridSearchCV

def find_best_mlp_model(train_x: pd.DataFrame, train_y: pd.Series):
    scale_pos_weight = (len(train_y) - train_y.sum()) / train_y.sum()

    parameter_space = {
    'hidden_layer_sizes': [(50, 50, 50), (100,50,50), (200,100,50), (50,100,50), ],
    'activation': ['tanh', 'relu', 'logistic'],
    'solver': ['sgd', 'adam'],
    'alpha': [0.0001, 0.05],
    'learning_rate': ['constant','adaptive'],
    'batch_size': [100, 200]
    }
    
    gsearch = GridSearchCV(
        estimator=MLPClassifier(
            max_iter=100,
        ),
        param_grid=parameter_space,
        scoring="roc_auc",
        n_jobs=32,
        cv=8,
    )

    gsearch.fit(train_x, train_y)

    return gsearch.best_params_, gsearch.best_score_

best_params, best_score = find_best_mlp_model(train_x, train_y)

best_params

best_score

mlp_model = MLPClassifier(
    hidden_layer_sizes=(50, 200, 50,),
    random_state=27,
    activation='logistic',
    solver="adam",
    max_iter=1000,
    batch_size=250,
    alpha=0.01,
    learning_rate="adaptive",
    learning_rate_init=1e-4,
    tol=1e-5
)

%%time

mlp_model.fit(train_x, train_y)

mlp_model.score(test_x, test_y)

from joblib import dump

dump(mlp_model, '../../models/mlp.joblib')
