==> notebooks/counterfactuals/02-datapoints.ipynb <==
import pandas as pd

df = pd.read_csv("../../datasets/FICO/heloc_dataset_v1.csv")

from xgboost.sklearn import XGBClassifier

model = XGBClassifier()

model.load_model("../../models/xgboost.json")

probabilities = [0.6 + i/50.0 for i in range(10)]
print(probabilities)

import utils

series = [utils.get_negative_closest(model, probability) for probability in probabilities]

df = pd.concat(series, axis=1).T

len(df.index.unique())

df

df.to_csv("../../datasets/datapoints.csv", index=False)
