==> notebooks/counterfactuals/03-trusty.ipynb <==
# TrustyAI

## Load data

import warnings
warnings.filterwarnings('ignore')

import pandas as pd

df = pd.read_csv("../../datasets/FICO/heloc_dataset_v1.csv")

import utils

bounds = utils.data_bounds(df)

## Initialise TrustyAI

import trustyai
import os
import site

DEFAULT_DEP_PATH = os.path.join(site.getsitepackages()[0], "trustyai", "dep")

CORE_DEPS = [
    f"../../deps/*",
    f"{DEFAULT_DEP_PATH}/org/slf4j/slf4j-api/1.7.30/slf4j-api-1.7.30.jar",
    f"{DEFAULT_DEP_PATH}/org/apache/commons/commons-lang3/3.12.0/commons-lang3-3.12.0.jar",
    f"{DEFAULT_DEP_PATH}/org/optaplanner/optaplanner-core-impl/8.18.0.Final/"
    f"optaplanner-core-impl-8.18.0.Final.jar",
    f"{DEFAULT_DEP_PATH}/org/apache/commons/commons-math3/3.6.1/commons-math3-3.6.1.jar",
    f"{DEFAULT_DEP_PATH}/org/kie/kie-api/8.18.0.Beta/kie-api-8.18.0.Beta.jar",
    f"{DEFAULT_DEP_PATH}/io/micrometer/micrometer-core/1.8.2/micrometer-core-1.8.2.jar",
]

trustyai.init(path=CORE_DEPS)

## XGBoost

### Load model

from xgboost.sklearn import XGBClassifier

model = XGBClassifier()

model.load_model("../../models/xgboost.json")

### Selecting inputs

Select the input $X$ with a probability of negative outcome closest to $75\%$.

datapoints = pd.read_csv("../../datasets/datapoints.csv")

datapoints

Check the model's output for $X$

X = datapoints.iloc[0]
model.predict_proba(X.to_numpy().reshape(1, -1))

TrustyAI wrapper for the XGBoost model

from org.kie.kogito.explainability.model import PredictionInput, PredictionOutput
from trustyai.model import output
import numpy as np

TARGET = "RiskPerformance"

def predict(_model):
    def fun(inputs):
        values = [_feature.value.as_number() for _feature in inputs[0].features]
        result = _model.predict_proba(np.array([values]))
        bad_prob, good_prob = result[0]
        if bad_prob > good_prob:
            _prediction = (0, bad_prob)
        else:
            _prediction = (1, good_prob)
        _output = output(name=TARGET, dtype="number", value=_prediction[0], score=_prediction[1])
        return [PredictionOutput([_output])]
    return fun

from trustyai.model import Model

provider = Model(predict(model))

Build input features with a fixed search bound of $[-20, 1000]$

from trustyai.model import feature

def build_input_feature(_X):
    input_feature = []
    input_dict = _X.to_dict()
    for name in input_dict:
        input_feature.append(
            feature(name=name, value=input_dict[name], dtype="number",
                    domain=(-20, 1000))
        )
    return input_feature

input_feature = build_input_feature(X)
predict(model)([PredictionInput(input_feature)])[0].outputs.get(0).toString()

Define the counterfactual goal as the label `1` ("Good" assessment)

goal = [output(name=TARGET, dtype="number", value=1)]

goal[0].toString()

from trustyai.model import counterfactual_prediction

prediction = counterfactual_prediction(
    input_features=input_feature,
    outputs=goal)

Create the explainer with a maximum of $100,000$ iterations.

from trustyai.explainers import CounterfactualExplainer

explainer = CounterfactualExplainer(steps=200_000)

Run the explainer

explanation = explainer.explain(prediction, provider)

Extract the features from the counterfactual explanation

expl_features = [e.asFeature() for e in explanation.entities]

cf_p = predict(model)([PredictionInput(expl_features)])[0].outputs.get(0)
cf_p = {
    "GoalValue": [cf_p.getValue().asNumber()],
    "GoalName": [str(cf_p.getName())],
    "GoalScore": [cf_p.getScore()],
}
print(cf_p)

def show_changes(explanation, original):
    entities = explanation.entities
    N = len(original)
    for i in range(N):
        name = original[i].name
        original_value = original[i].value.as_number()
        new_value = entities[i].as_feature().value.as_number()
        if original_value != new_value:
            print(f"Feature '{name}': {original_value} -> {new_value}")


show_changes(explanation, input_feature)

### Run for all datapoints

import utils

counterfactuals = []
for i in range(10):
    print(f"Running counterfactual search #{i}")
    X = datapoints.iloc[i]
    input_feature = build_input_feature(X)
    prediction = counterfactual_prediction(input_features=input_feature, outputs=goal)
    explanation = explainer.explain(prediction, provider)
    expl_features = [e.asFeature() for e in explanation.entities]
    cf_p = predict(model)([PredictionInput(expl_features)])[0].outputs.get(0)
    cf_p = {
        "GoalValue": [cf_p.getValue().asNumber()],
        "GoalName": [str(cf_p.getName())],
        "GoalScore": [cf_p.getScore()],
    }
    d_cf = {f"Cf{f.name}": [f.value.as_number()] for f in expl_features}
    cf_df = utils.save_result(
        original=X, cf=d_cf, score=cf_p, method=["TrustyAI"], model=["XGBoost"]
    )
    counterfactuals.append(cf_df)

xgboost_final = pd.concat(counterfactuals)

## MLP/Adam

### Load model

from joblib import load

mlp_model = load('../../models/mlp.joblib') 

mlp_predict = predict(mlp_model)

mlp_predict([PredictionInput(input_feature)])[0].outputs.get(0).toString()

mlp_model.predict_proba(X.to_numpy().reshape(1, -1))

mlp_provider = Model(mlp_predict)

import utils

counterfactuals = []
for i in range(10):
    print(f"Running counterfactual search #{i}")
    X = datapoints.iloc[i]
    input_feature = build_input_feature(X)
    prediction = counterfactual_prediction(input_features=input_feature, outputs=goal)
    explanation = explainer.explain(prediction, mlp_provider)
    expl_features = [e.asFeature() for e in explanation.entities]
    cf_p = predict(model)([PredictionInput(expl_features)])[0].outputs.get(0)
    cf_p = {
        "GoalValue": [cf_p.getValue().asNumber()],
        "GoalName": [str(cf_p.getName())],
        "GoalScore": [cf_p.getScore()],
    }
    d_cf = {f"Cf{f.name}": [f.value.as_number()] for f in expl_features}
    cf_df = utils.save_result(
        original=X, cf=d_cf, score=cf_p, method=["TrustyAI"], model=["MLP"]
    )
    counterfactuals.append(cf_df)

mlp_final = pd.concat(counterfactuals)

final_df = pd.concat([xgboost_final, mlp_final])

final_df.T

final_df.to_csv("../../results/cf-trustyai.csv")
