==> notebooks/counterfactuals/04a-alibi-proto.ipynb <==
# ALIBI-proto
## Load data

import warnings
warnings.filterwarnings('ignore')

import pandas as pd

df = pd.read_csv("../../datasets/FICO/heloc_dataset_v1.csv")

TARGET = "RiskPerformance"

from sklearn.model_selection import train_test_split

df[TARGET] = df[TARGET].factorize()[0]

train, test = train_test_split(df, test_size=0.25, random_state=42)

train_x = train[df.columns[~train.columns.isin([TARGET])]]

## XGBoost

### Load model

from xgboost.sklearn import XGBClassifier

model = XGBClassifier()

model.load_model("../../models/xgboost.json")

### Standard counterfactuals

import tensorflow as tf

tf.keras.backend.clear_session()
tf.compat.v1.disable_eager_execution()
tf.compat.v1.disable_v2_behavior()

from alibi.explainers import CounterfactualProto

shape = (1, 23)

predict_fn = lambda x: model.predict_proba(x)

def build_cf(pfn):
    _cf = CounterfactualProto(
        pfn,
        shape,
        use_kdtree=True,
        theta=100.0,
        max_iterations=200,
        feature_range=(-20, 1000),
        c_init=2.0,
        c_steps=80,
    )
    return _cf

cf = build_cf(predict_fn)

import utils
import pandas as pd

datapoints = pd.read_csv("../../datasets/datapoints.csv")

cf.fit(train_x.to_numpy())

### Store results

def create_result_df(_explanation, X, model: str):
    N_cf = 1
    schema = utils.get_result_schema()
    v = [[] for i in range(23)]
    o = [[] for i in range(23)]
    goalValue = []
    goalScore = []
    for i in range(N_cf):
        ccf = _explanation['data']['cf']
        values = ccf['X'].tolist()[0]
        goalValue.append(ccf['class'])
        goalScore.append(ccf['proba'].tolist()[0][1])
        for j in range(23):
            v[j].append(values[j])
            o[j].append(X[0][j])

    d = {e[0]: e[1] for e in zip(schema[23:-4], v)}
    od = {e[0]: e[1] for e in zip(schema[:23], o)}
    goal = {"GoalValue": goalValue, "GoalScore": goalScore, "GoalName": ["RiskPerformance"] * N_cf}
    method = ["Alibi-Proto"] * N_cf

    return utils.save_result(original=od, cf=d, score=goal, method=method, model=model)

counterfactuals = []
for i in range(10):
    print(f"Running counterfactual search #{i}")
    X = datapoints.iloc[i].to_numpy().reshape(1, -1)
    explanation = cf.explain(X)
    result = create_result_df(explanation, X, "XGBoost")
    counterfactuals.append(result)

xgboost_final = pd.concat(counterfactuals)

xgboost_final

## MLP

## Load model

from joblib import load

mlp_model = load('../../models/mlp.joblib') 

mlp_predict_fn = lambda x: mlp_model.predict_proba(x)

mlp_cf = build_cf(mlp_predict_fn)

mlp_cf.fit(train_x.to_numpy())

counterfactuals = []
for i in range(10):
    print(f"Running counterfactual search #{i}")
    X = datapoints.iloc[i].to_numpy().reshape(1, -1)
    explanation = mlp_cf.explain(X)
    result = create_result_df(explanation, X, "MLP")
    counterfactuals.append(result)

mlp_final = pd.concat(counterfactuals)

final_df = pd.concat([xgboost_final, mlp_final])

final_df

final_df.to_csv("../../results/cf-alibi-proto.csv")
